requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==76.0.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
